+++ src/openrct2/actions/ClearAction.h
@@ -9,6 +9,7 @@
 
 #pragma once
 
+#include "../OpenRCT2.h"
 #include "../management/Finance.h"
 #include "GameAction.h"
 
@@ -45,6 +46,35 @@ private:
     GameActions::Result::Ptr CreateResult() const;
     GameActions::Result::Ptr QueryExecute(bool executing) const;
     money32 ClearSceneryFromTile(const CoordsXY& tilePos, bool executing) const;
-    static void ResetClearLargeSceneryFlag();
-    static bool MapCanClearAt(const CoordsXY& location);
+
+    /**
+     * Function to clear the flag that is set to prevent cost duplication
+     * when using the clear scenery tool with large scenery.
+     */
+    static void ResetClearLargeSceneryFlag()
+    {
+        // TODO: Improve efficiency of this
+        for (int32_t y = 0; y < MAXIMUM_MAP_SIZE_TECHNICAL; y++)
+        {
+            for (int32_t x = 0; x < MAXIMUM_MAP_SIZE_TECHNICAL; x++)
+            {
+                auto tileElement = map_get_first_element_at(TileCoordsXY{ x, y }.ToCoordsXY());
+                do
+                {
+                    if (tileElement == nullptr)
+                        break;
+                    if (tileElement->GetType() == TILE_ELEMENT_TYPE_LARGE_SCENERY)
+                    {
+                        tileElement->AsLargeScenery()->SetIsAccounted(false);
+                    }
+                } while (!(tileElement++)->IsLastForTile());
+            }
+        }
+    }
+
+    static bool MapCanClearAt(const CoordsXY& location)
+    {
+        return (gScreenFlags & SCREEN_FLAGS_SCENARIO_EDITOR) || gCheatsSandboxMode
+            || map_is_location_owned_or_has_rights(location);
+    }
 };
