+++ src/openrct2/actions/RideEntranceExitPlaceAction.h
@@ -9,6 +9,7 @@
 
 #pragma once
 
+#include "../world/Entrance.h"
 #include "GameAction.h"
 
 DEFINE_GAME_ACTION(RideEntranceExitPlaceAction, GAME_COMMAND_PLACE_RIDE_ENTRANCE_OR_EXIT, GameActions::Result)
@@ -45,4 +46,42 @@ public:
     GameActions::Result::Ptr Query() const override;
 
     GameActions::Result::Ptr Execute() const override;
+
+    static GameActions::Result::Ptr TrackPlaceQuery(const CoordsXYZ& loc, const bool isExit)
+    {
+        auto errorTitle = isExit ? STR_CANT_BUILD_MOVE_EXIT_FOR_THIS_RIDE_ATTRACTION
+                                 : STR_CANT_BUILD_MOVE_ENTRANCE_FOR_THIS_RIDE_ATTRACTION;
+        if (!map_check_free_elements_and_reorganise(1))
+        {
+            return MakeResult(GameActions::Status::NoFreeElements, errorTitle);
+        }
+
+        if (!gCheatsSandboxMode && !map_is_location_owned(loc))
+        {
+            return MakeResult(GameActions::Status::NotOwned, errorTitle);
+        }
+
+        int16_t baseZ = loc.z;
+        int16_t clearZ = baseZ + (isExit ? RideExitHeight : RideEntranceHeight);
+        auto cost = MONEY32_UNDEFINED;
+        if (!map_can_construct_with_clear_at(
+                { loc, baseZ, clearZ }, &map_place_non_scenery_clear_func, { 0b1111, 0 }, 0, &cost, CREATE_CROSSING_MODE_NONE))
+        {
+            return MakeResult(GameActions::Status::NoClearance, errorTitle, gGameCommandErrorText, gCommonFormatArgs);
+        }
+
+        if (gMapGroundFlags & ELEMENT_IS_UNDERWATER)
+        {
+            return MakeResult(GameActions::Status::Disallowed, errorTitle, STR_RIDE_CANT_BUILD_THIS_UNDERWATER);
+        }
+
+        if (baseZ > MaxRideEntranceOrExitHeight)
+        {
+            return MakeResult(GameActions::Status::Disallowed, errorTitle, STR_TOO_HIGH);
+        }
+        auto res = MakeResult();
+        res->Position = { loc.ToTileCentre(), tile_element_height(loc) };
+        res->Expenditure = ExpenditureType::RideConstruction;
+        return res;
+    }
 };
