+++ src/openrct2/actions/WaterLowerAction.cpp
@@ -9,6 +9,7 @@
 
 #include "WaterLowerAction.h"
 
+#include "../OpenRCT2.h"
 #include "../audio/audio.h"
 #include "WaterSetHeightAction.h"
