+++ src/openrct2/libopenrct2.vcxproj
@@ -461,12 +461,88 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="actions\BalloonPressAction.cpp" />
+    <ClCompile Include="actions\BannerPlaceAction.cpp" />
+    <ClCompile Include="actions\BannerRemoveAction.cpp" />
+    <ClCompile Include="actions\BannerSetColourAction.cpp" />
+    <ClCompile Include="actions\BannerSetNameAction.cpp" />
+    <ClCompile Include="actions\BannerSetStyleAction.cpp" />
+    <ClCompile Include="actions\ClearAction.cpp" />
+    <ClCompile Include="actions\ClimateSetAction.cpp" />
+    <ClCompile Include="actions\CustomAction.cpp" />
+    <ClCompile Include="actions\FootpathAdditionPlaceAction.cpp" />
+    <ClCompile Include="actions\FootpathAdditionRemoveAction.cpp" />
+    <ClCompile Include="actions\FootpathPlaceAction.cpp" />
+    <ClCompile Include="actions\FootpathPlaceFromTrackAction.cpp" />
+    <ClCompile Include="actions\FootpathRemoveAction.cpp" />
     <ClCompile Include="actions\GameAction.cpp" />
     <ClCompile Include="actions\GameActionCompat.cpp" />
     <ClCompile Include="actions\GameActionRegistration.cpp" />
+    <ClCompile Include="actions\GuestSetFlagsAction.cpp" />
+    <ClCompile Include="actions\GuestSetNameAction.cpp" />
+    <ClCompile Include="actions\LandBuyRightsAction.cpp" />
+    <ClCompile Include="actions\LandLowerAction.cpp" />
+    <ClCompile Include="actions\LandRaiseAction.cpp" />
+    <ClCompile Include="actions\LandSetHeightAction.cpp" />
+    <ClCompile Include="actions\LandSetRightsAction.cpp" />
+    <ClCompile Include="actions\LandSmoothAction.cpp" />
+    <ClCompile Include="actions\LargeSceneryPlaceAction.cpp" />
+    <ClCompile Include="actions\LargeSceneryRemoveAction.cpp" />
+    <ClCompile Include="actions\LargeScenerySetColourAction.cpp" />
+    <ClCompile Include="actions\LoadOrQuitAction.cpp" />
+    <ClCompile Include="actions\MazePlaceTrackAction.cpp" />
+    <ClCompile Include="actions\MazeSetTrackAction.cpp" />
+    <ClCompile Include="actions\NetworkModifyGroupAction.cpp" />
+    <ClCompile Include="actions\ParkEntranceRemoveAction.cpp" />
+    <ClCompile Include="actions\ParkMarketingAction.cpp" />
+    <ClCompile Include="actions\ParkSetDateAction.cpp" />
+    <ClCompile Include="actions\ParkSetLoanAction.cpp" />
+    <ClCompile Include="actions\ParkSetNameAction.cpp" />
+    <ClCompile Include="actions\ParkSetParameterAction.cpp" />
+    <ClCompile Include="actions\ParkSetResearchFundingAction.cpp" />
+    <ClCompile Include="actions\PauseToggleAction.cpp" />
+    <ClCompile Include="actions\PeepPickupAction.cpp" />
+    <ClCompile Include="actions\PlaceParkEntranceAction.cpp" />
+    <ClCompile Include="actions\PlacePeepSpawnAction.cpp" />
+    <ClCompile Include="actions\PlayerKickAction.cpp" />
+    <ClCompile Include="actions\PlayerSetGroupAction.cpp" />
+    <ClCompile Include="actions\RideCreateAction.cpp" />
+    <ClCompile Include="actions\RideDemolishAction.cpp" />
+    <ClCompile Include="actions\RideEntranceExitPlaceAction.cpp" />
+    <ClCompile Include="actions\RideEntranceExitRemoveAction.cpp" />
+    <ClCompile Include="actions\RideSetAppearanceAction.cpp" />
+    <ClCompile Include="actions\RideSetColourSchemeAction.cpp" />
+    <ClCompile Include="actions\RideSetNameAction.cpp" />
+    <ClCompile Include="actions\RideSetPriceAction.cpp" />
+    <ClCompile Include="actions\RideSetSettingAction.cpp" />
+    <ClCompile Include="actions\RideSetStatusAction.cpp" />
+    <ClCompile Include="actions\RideSetVehicleAction.cpp" />
+    <ClCompile Include="actions\ScenarioSetSettingAction.cpp" />
+    <ClCompile Include="actions\SetCheatAction.cpp" />
+    <ClCompile Include="actions\SetParkEntranceFeeAction.cpp" />
+    <ClCompile Include="actions\SignSetNameAction.cpp" />
+    <ClCompile Include="actions\SignSetStyleAction.cpp" />
     <ClCompile Include="actions\SmallSceneryPlaceAction.cpp" />
     <ClCompile Include="actions\SmallSceneryRemoveAction.cpp" />
+    <ClCompile Include="actions\SmallScenerySetColourAction.cpp" />
+    <ClCompile Include="actions\StaffFireAction.cpp" />
+    <ClCompile Include="actions\StaffHireNewAction.cpp" />
+    <ClCompile Include="actions\StaffSetColourAction.cpp" />
+    <ClCompile Include="actions\StaffSetCostumeAction.cpp" />
+    <ClCompile Include="actions\StaffSetNameAction.cpp" />
+    <ClCompile Include="actions\StaffSetOrdersAction.cpp" />
+    <ClCompile Include="actions\StaffSetPatrolAreaAction.cpp" />
+    <ClCompile Include="actions\SurfaceSetStyleAction.cpp" />
+    <ClCompile Include="actions\TileModifyAction.cpp" />
     <ClCompile Include="actions\TrackDesignAction.cpp" />
+    <ClCompile Include="actions\TrackPlaceAction.cpp" />
+    <ClCompile Include="actions\TrackRemoveAction.cpp" />
+    <ClCompile Include="actions\TrackSetBrakeSpeedAction.cpp" />
+    <ClCompile Include="actions\WallPlaceAction.cpp" />
+    <ClCompile Include="actions\WallRemoveAction.cpp" />
+    <ClCompile Include="actions\WallSetColourAction.cpp" />
+    <ClCompile Include="actions\WaterLowerAction.cpp" />
+    <ClCompile Include="actions\WaterRaiseAction.cpp" />
+    <ClCompile Include="actions\WaterSetHeightAction.cpp" />
     <ClCompile Include="audio\Audio.cpp" />
     <ClCompile Include="audio\AudioMixer.cpp" />
     <ClCompile Include="audio\DummyAudioContext.cpp" />
