+++ src/openrct2/peep/Staff.cpp
@@ -31,6 +31,7 @@
 #include "../ride/Track.h"
 #include "../scenario/Scenario.h"
 #include "../util/Util.h"
+#include "../windows/Intent.h"
 #include "../world/Entrance.h"
 #include "../world/Footpath.h"
 #include "../world/Scenery.h"
